move every publicly accessible class to the api package
