+++ flink/v1.20/flink/src/main/java/org/apache/iceberg/flink/maintenance/api/TaskResult.java
@@ -16,15 +16,13 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iceberg.flink.maintenance.operator;
+package org.apache.iceberg.flink.maintenance.api;
 
 import java.io.Serializable;
 import java.util.List;
-import org.apache.flink.annotation.Internal;
 import org.apache.iceberg.relocated.com.google.common.base.MoreObjects;
 
 /** The result of a single Maintenance Task. */
-@Internal
 public class TaskResult implements Serializable {
   private final int taskIndex;
   private final long startEpoch;
