+++ flink/v1.20/flink/src/main/java/org/apache/iceberg/flink/maintenance/operator/TableChange.java
@@ -37,7 +37,7 @@ public class TableChange {
   private long eqDeleteRecordCount;
   private int commitCount;
 
-  TableChange(
+  private TableChange(
       int dataFileCount,
       long dataFileSizeInBytes,
       int posDeleteFileCount,
